Fix WebSocket frame length truncation, send-error logging, and WS route constraints
- send_frame: 127-encoding stored htobe64() result in a uint16_t and then
  appended 8 bytes from a 2-byte variable (UB + broken lengths > 64KB)
- http server: 101 handshake failure logged res_exp.error() (a moved-from
  dispatch result) instead of the actual send error
- route(path, ws_handler): constraint required an invocable-with-no-args
  handler, while the real signature takes websocket::details::Connection&,
  making every WebSocket route registration fail to compile

+++ netx/websocket/connection.hpp
@@ -136,7 +136,7 @@ inline core::Task<core::Expected<>> Connection::send_frame(
 	else
 	{
 		header.push_back(127);
-		std::uint16_t len = htobe64(payload.size());
+		std::uint64_t len = htobe64(payload.size());
 		header.append(reinterpret_cast<const char*>(&len), 8);
 	}
